// apps/sales-copilot/src/generated/models/Crf5c_prompttemplatesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf5c_prompttemplatesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf5c_prompttemplatesstatecode = keyof typeof Crf5c_prompttemplatesstatecode;
export const Crf5c_prompttemplatesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf5c_prompttemplatesstatuscode = keyof typeof Crf5c_prompttemplatesstatuscode;

export interface Crf5c_prompttemplatesBase {
  crf5c_body?: string;
  crf5c_contractversion?: number;
  crf5c_description?: string;
  crf5c_ispublished?: boolean;
  crf5c_modeltier?: string;
  crf5c_name?: string;
  crf5c_notes?: string;
  crf5c_prompttemplateid: string;
  crf5c_promptversion?: number;
  crf5c_responseformat?: string;
  crf5c_variables?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Crf5c_prompttemplatesstatecode;
  statuscode?: Crf5c_prompttemplatesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crf5c_prompttemplates extends Crf5c_prompttemplatesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf5c_ispublishedname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}